+++ src/subscription/subscription.controller.ts
@@ -1,3 +1,4 @@
+import { UnsubscribeService } from './../notifications/v2/unsubscribe/unsubscribe.service';
 import {
     Body,
     Controller,
@@ -5,6 +6,7 @@ import {
     Get,
     Param,
     Post,
+    Query,
     Res,
 } from '@nestjs/common';
 import { Response } from 'express';
@@ -14,7 +16,10 @@ import { SubscriptionService } from './subscription.service';
 
 @Controller('subscriptions')
 export class SubscriptionController {
-    constructor(private subscriptionService: SubscriptionService) {}
+    constructor(
+        private subscriptionService: SubscriptionService,
+        private unsubscribeService: UnsubscribeService,
+    ) {}
     @Post()
     async create(@Body() dto: CreateSubscriptionDto): Promise<Subscription> {
         const result = await this.subscriptionService.create(dto);
@@ -22,38 +27,69 @@ export class SubscriptionController {
         return result;
     }
 
-    @Get('users/:plainTextEmailAddress')
-    async findAllByEmailAddress(
-        @Param('plainTextEmailAddress') plainTextEmailAddress: string,
+    @Get('users/:id')
+    async findAllBySubOrEmailAddress(
+        @Param('id') id: string,
     ): Promise<Subscription[]> {
-        return this.subscriptionService.findAllByEmailAddress(
-            plainTextEmailAddress,
-        );
+        return this.subscriptionService.findAllBySubOrEmailAddress(id);
     }
 
-    @Get('users/:plainTextEmailAddress/grants/:grantId')
-    async findByEmailAddressAndGrantId(
-        @Param('plainTextEmailAddress') plainTextEmailAddress: string,
+    @Get('users/:id/grants/:grantId')
+    async findByUserAndGrantId(
+        @Param('id') id: string,
         @Param('grantId') grantId: string,
     ): Promise<Subscription> {
-        return this.subscriptionService.findByEmailAndGrantId(
-            plainTextEmailAddress,
+        const subscription = await this.subscriptionService.findBySubAndGrantId(
+            id,
             grantId,
         );
+
+        return subscription
+            ? subscription
+            : this.subscriptionService.findByEmailAndGrantId(id, grantId);
     }
 
-    @Delete('users/:plainTextEmailAddress/grants/:grantId')
-    async deleteByEmailAndGrantId(
-        @Param('plainTextEmailAddress') plainTextEmailAddress: string,
+    @Delete('users/:id/grants/:grantId')
+    async deleteByUserAndGrantId(
+        @Param('id') id: string,
         @Param('grantId') grantId: string,
         @Res() response: Response,
+        @Query() query: { unsubscribeReference?: string },
     ): Promise<void> {
-        const result = await this.subscriptionService.deleteByEmailAndGrantId(
-            plainTextEmailAddress,
+        const ref = query?.unsubscribeReference;
+        let result = await this.subscriptionService.deleteBySubAndGrantId(
+            id,
             grantId,
         );
+        if (result.affected == 0) {
+            result = await this.subscriptionService.deleteByEmailAndGrantId(
+                id,
+                grantId,
+            );
+        }
 
-        result.affected == 0 ? response.status(404) : response.status(204);
-        response.send();
+        if (ref && ref !== 'undefined') {
+            await this.unsubscribeService
+                .deleteOneById(ref)
+                .catch((error: unknown) => {
+                    console.error(
+                        `Failed to unsubscribe from unsubscribeReference: ${ref}.
+                        error:${JSON.stringify(error)}`,
+                    );
+                });
+        } else {
+            await this.unsubscribeService
+                .deleteOneBySubOrEmail(id, { subscriptionId: grantId })
+                .catch((error: unknown) => {
+                    console.error(
+                        `Failed to unsubscribe from sub: ${id}. error:${JSON.stringify(
+                            error,
+                        )}`,
+                    );
+                });
+        }
+
+        response.status(result.affected == 0 ? 404 : 204);
+        response.end();
     }
 }
